Implement save_grant
save_grant creates a grant object for the specified client and code and
inserts it to auth_db. Because this is using insert, it will fail if an
existing grant for the code already exists.

The document ttl is currently hard-coded to 600, indicating that the
document (i.e.: the grant) expires in 5 minutes. This value is meant
for temporary development use, only.