Removes Flask in favour of Express
This project will be difficult enough to finish on my own, so I’m going
to reduce the number of technologies to learn. This API was cooked up
just to make progress on the front end, so I’ve split it out from the
front end and put it here where it belongs.