Fixes bugs with existing password validation and user verification
User verification was getting assigned a value because it was
evaluating false the same as an empty string in the user constructor.
Switched to using a true value and evaluating explicitly for a boolean.

Existing password validation failed on password change because it was
testing the existing password provided against the new password
submitted. This has been corrected.