Fixes CORs handling
Chrome’s preflight options request was leading to a 404 and it wouldn’t
allow request to proceed, even with the valid CORs headers that are
attached to the response. Other browsers didn’t seem to care. But now
will a return a 200 OK to any OPTIONS request before getting to
anything else.